requirements.txt
fastapi
uvicorn
gunicorn
sentence-transformers
faiss-cpu
pydantic
numpy
python-multipart
